data_scorer/requirements.txt
torch==2.6.0
datasets==3.6.0
vllm==0.8.4
pyyaml
tiktoken==0.8.0
tqdm==4.67.0
pandas==1.5.3
numpy==1.26.3
sentence-transformers==3.3.1
nltk==3.9.1
langdetect==1.0.9
benepar==0.2.0
fasttext==0.9.3
openai==1.55.0
prettytable==3.16.0
pydantic
rich
typer
aiofiles
seaborn
matplotlib
orjson 
pyarrow==20.0.0
datasketch==1.7.0
faiss-gpu==1.7.2
lexicalrichness==0.5.1
